+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - Kokkos now compiles and links separately to fix complication with Kokkos 2.05.00.
+- Kokkos can now instantiate single and double precision.
 - OpenMP 4.5 map and reduction clause order to ensure reduction result copied back.